// lib/db/forms-db.types.tsx
/**
 * @ Author: BetterInternship
 * @ Create Time: 2025-10-11 00:00:00
 * @ Modified time: 2026-04-18 23:50:39
 * @ Description:
 *
 * This handles interactions with our MOA Api server.
 */

export interface FormTemplate {
  formDocument: string;
  formVersion: number;
  formName: string;
  formLabel: string;
}