// lib/api/generated/models/createJobChallengeListingDto.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * BetterInternship API
 * The official API of the BetterInternship website.
 * OpenAPI spec version: 2.0
 */
import type { CreateJobChallengeListingDtoInternshipPreferences } from "./createJobChallengeListingDtoInternshipPreferences";
import type { CreateJobChallengeDto } from "./createJobChallengeDto";

export interface CreateJobChallengeListingDto {
  /** @nullable */
  allowance: number | null;
  description: string;
  internship_preferences?: CreateJobChallengeListingDtoInternshipPreferences;
  /** @nullable */
  is_active: boolean | null;
  is_unlisted: boolean;
  /** @nullable */
  location: string | null;
  /** @nullable */
  requirements: string | null;
  /** @nullable */
  salary: string | null;
  /** @nullable */
  salary_freq: number | null;
  title: string;
  challenge: CreateJobChallengeDto;
}