// lib/api/generated/models/registerUserDto.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * BetterInternship API
 * The official API of the BetterInternship website.
 * OpenAPI spec version: 2.0
 */

export interface RegisterUserDto {
  university: string;
  degree?: string;
  first_name?: string;
  middle_name?: string;
  last_name?: string;
}